+++ lambda-callback.tf
@@ -111,9 +111,10 @@ resource "aws_lambda_function" "callback" {
 
   function_name = "${module.labels.id}-callback"
   handler       = "callback.handler"
+  layers        = lookup(var.lambda_custom_runtimes, "callback", "NOT-FOUND") == "NOT-FOUND" ? null : var.lambda_custom_runtimes["callback"].layers
   memory_size   = var.lambda_callback_memory_size
   role          = aws_iam_role.callback.arn
-  runtime       = "nodejs12.x"
+  runtime       = lookup(var.lambda_custom_runtimes, "callback", "NOT-FOUND") == "NOT-FOUND" ? var.lambda_default_runtime : var.lambda_custom_runtimes["callback"].runtime
   tags          = module.labels.tags
   timeout       = var.lambda_callback_timeout