+++ lambda-cso.tf
@@ -87,9 +87,10 @@ resource "aws_lambda_function" "cso" {
 
   function_name = "${module.labels.id}-cso"
   handler       = "cso.handler"
+  layers        = lookup(var.lambda_custom_runtimes, "cso", "NOT-FOUND") == "NOT-FOUND" ? null : var.lambda_custom_runtimes["cso"].layers
   memory_size   = var.lambda_cso_memory_size
   role          = aws_iam_role.cso[0].arn
-  runtime       = "nodejs12.x"
+  runtime       = lookup(var.lambda_custom_runtimes, "cso", "NOT-FOUND") == "NOT-FOUND" ? var.lambda_default_runtime : var.lambda_custom_runtimes["cso"].runtime
   tags          = module.labels.tags
   timeout       = var.lambda_cso_timeout