+++ lambda-settings.tf
@@ -82,9 +82,10 @@ resource "aws_lambda_function" "settings" {
 
   function_name = "${module.labels.id}-settings"
   handler       = "settings.handler"
+  layers        = lookup(var.lambda_custom_runtimes, "settings", "NOT-FOUND") == "NOT-FOUND" ? null : var.lambda_custom_runtimes["settings"].layers
   memory_size   = var.lambda_settings_memory_size
   role          = aws_iam_role.settings.arn
-  runtime       = "nodejs12.x"
+  runtime       = lookup(var.lambda_custom_runtimes, "settings", "NOT-FOUND") == "NOT-FOUND" ? var.lambda_default_runtime : var.lambda_custom_runtimes["settings"].runtime
   tags          = module.labels.tags
   timeout       = var.lambda_settings_timeout